[feature] hw_L65: created 'products.routes.js', replaced products routes from 'server.js' to 'products.routes.js', changes applied
